+++ src/__generated__/optionsType.ts
@@ -123,6 +123,12 @@ export type BibTeXTidyOptions = {
 	 */
 	escape?: boolean | 'new';
 	/**
+	 * Unescape special characters
+	 *
+	 * Convert LaTeX escapes to Unicode characters.
+	 */
+	unescape?: boolean;
+	/**
 	 * Sort fields
 	 *
 	 * Sort the fields within entries.